Change WAIT_FOR_AUTO_NEG_DEFAULT to 0.  Do not wait for auto-negotiation to complete before continuing.  if_em can take over two seconds to autoneg, and it renegotiates virtually every time the interface is manipulated.  For example, running dhclient will cause if_em to renegotiate 3 times in a row.
Note that even when we were waiting for autoneg, the interface would not
always be completely up after the autoneg completed.  This is possibly
due to the watchdog timer (which is on a 2 second interval).  In fact, on
my test box this is what was tickling the recently fixed DNS bug (timing
out on part of the dns search path) and subsequent NFS mount failures.