Fix a bug in sys/pipe.c.  xio_init_ubuf() might not be able to load up the requested number of bytes even if the request is limited to XIO_INTERNAL_SIZE if the user buffer base is not page-aligned.  XIO will set xio_bytes to the actual size of the buffer.
Note that this bug was never exercised due to the 64KB pipe kmem buffer size
limit, so it could not have been the cause of recent problems.

Use kmem_alloc_nofault() instead of kmem_alloc_pageable() for the kmem
reservation.  This is more correct but should have no actual effect on
the system.