Document the ``resource management'' (rman) abstraction in rman(9).
The rman(9) manual page has sufficient information on the rman_*
functions and some background on the concept of regions and
resources and how rman(9) abstracts them.

Install the necessary MLINKs.

Note:
Some of the functions have been commented out in the mdoc(7) until
I or Joerg get around to updating the DragonFly rman(9) abstraction.
After that is done, other parts of the maual page can be expanded.

Obtained-from:  	FreeBSD (HEAD)
Adapted-by:     	Hiten Pandya <hmp@backplane.com>