Make sysctl_kern_proc() iterate through available cpus to retrieve the list of threads installed on each one rather then just retrieve the threads on the cpu it happens to have been called from.
To accomplish this, sysctl_kern_proc() now uses the new LWKT API to
migrate itself to each target cpu in order to safely access the thread
list on that cpu, ending up on the same cpu it started from.  The overhead
is nasty to be sure, but who cares when only 'top' and 'ps' and related
non-critical utilities use the feature?