Use a normal malloc() for PCB allocations instead of the really aweful mbuf hacks that were previously being used (which also gets rid of the remaining dtom() calls).
Get rid of the packet/queue overloading and instead malloc() a queue header
for queueing packets.  The old code depended on dtom() to figure out the
mbuf associated with the packet and we are removing dtom().

Give NS a default usrreq so it doesn't panic the system on boot, but note
that NS will remain broken until the usrreq can be tied into the existing
code.

Review and testing by: Hiten Pandya and David Rhodus