C99 update: freopen(3) with NULL 'path' argument so that it opens the same file with a different mode.
At the moment, we just recycle the old file descriptor instead of storing
it somewhere and using that to reopen the file, as the standard seems to
require.  Strictly conforming C99 applications would probably not be able
to tell the difference but POSIX ones can.

While I am there, update the manual page about the new functionality and
correct nearby mdoc(7) markup errors.

Original-work-by:    	Tim J. Robbins <tjr@freebsd.org>
Additional-work-by: 	Hiten Pandya <hmp@backplane.com>