Move the 'p_start' field from struct pstats (Process Statistics) into the thread structure and call it 'td_start'.  The behavior of vm_fork(9) is retained, i.e., it still copies the start time from the parent process just as it did before.
The 'td_start' will later be used by pure threads to indicate their start
time.  It has not been committed in this round because use of the microtime()
function at such a early point in the boot process might be unsafe.

Note, there should be no problem in accessing the td_start field, unless
the process is a Zombie; due to the way Zombies are reaped, the thread will
be decoupled in kern_wait1() but the process will still be around for a
while it will not be possible to access the td_start field in such
scenarios.  A little note about this has been added on top of struct proc
in <sys/proc.h> for future reference.

This work was a collaboration of Hiten Pandya <hmp@backplane.com> and
Matthew Dillon <dillon@apollo.backplane.com>