Bring in 1.33 from FreeBSD-5:
When rebooting the machine jump to 0xf000:0xfff0 instead of 0xffff:0x0.
While we end up the same place, we end up with two different CS register
values after the jump and 0xf000 is compatible with the hardware reset value.
This makes a difference if the BIOS does a near jump before a far jump.

Detective work and patch by:     Adrian Steinmann <ast@marabu.ch>