Fix a minor bug in the auto-console selection (handle the -m mute option properly).
Detect when the serial port is unmapped (reads 0xFF) by putting a limit in
the flush code and returning a testable value, then refuse to automatically
configure it if so because an unmapped serial port always looks like it has
input pending.

This fixes a dual-console issue with laptops that don't map the serial port
during the boot stage (e.g. my emachines amd64 laptop).

Be ultra conservative for now, do not try to initialize the FIFO.