ata-raid associates raw ata disk devices to record the raid setup and checks for the existance of an associated si_disk structure to determine whether the associations remain valid.  In DragonFly the abstracted 'disk' layer is now above the raw drive device rather then integrated with the raw drive device and thus the raw device does not have an si_disk association.  Change the conditionals to simply check for the existance of the device.
The prior check was resulting in drives being marked damaged in the raid
when they weren't, making the raid unusable and requiring a complete rebuild.
Though it is possible to rebuild and regenerate the labels (e.g. using
scan_ffs) without losing data, it is painfull to have to do it.

Reported-painfully-by: Andrew Atrens <atrens@nortelnetworks.com>