VFS messaging/interfacing work stage 5/99.  Start work on the new namecache/lookup API.  The core of this API will be the concept of a 'locked namespace' rather then a 'locked vnode' for namespace centric operations like CREATE, DELETE, and RENAME.   The namecache will no longer be optional for a VFS but instead will become centric to all kernel namespace operations.
This commit is mostly non-functional.  It removes an extension I had made
to VOP_LOOKUP/VOP_CACHEDLOOKUP and starts adding support functions for the
new cache_nclookup() API.  The work being backed out was originally intended
to shepard the new work, but the new interface is now so different (and far
more simplified) then the original that it's actually better to create a
new VOP for it instead of augmenting existing VOPs.