Fix bugs in the limiting code for negative-hit namecache entries.  The system was allowing too many negative entries to build up.  Since these entries by definition do not have associated vnodes the vnode code is not responsible for limiting the number of negative entries.  Since vnodes associated with non-leaf namecache entries are held, the build-up of negative entries prevented directory vnodes from being recycled and caused vnlru_proc to fail.
To fix this we check for and cleanup negative entries in both the old API
cache_enter() and the new API cache_nlookup().  Rather then just removing one
when we exceed the limit we remove 10 and we create a 90%/100% hysteresis
to avoid a non-optimal following cleanup call on every cache lookup.

Reported-by: Pawel Biernacki <kaktus@DragonFlyBSD.pl>,
	     Chris Pressey <cpressey@catseye.mine.nu>
Testing-by: same